requirements.txt
Django==5.2.5
djangorestframework==3.16.1
django-cors-headers==4.7.0
gunicorn==23.0.0
psycopg2-binary==2.9.10
djangorestframework-simplejwt==5.3.0
django-filter==23.5
whitenoise==6.6.0
